requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
asyncmy==0.2.7
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
discord.py==2.2.2
distlib==0.3.6
fastapi==0.95.0
filelock==3.9.0
fonttools==4.39.3
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
identify==2.5.11
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numpy==1.24.2
osrsbox==2.2.3
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
pre-commit==2.21.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.2
PyYAML==6.0
requests==2.28.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.26.1
tomli==2.0.1
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
virtualenv==20.17.1
watchfiles==0.19.0
yarl==1.8.2
